Make taxBureau aggregate return a basic system report

+++ src/main/java/com/anzisolutions/bankingsimulator/bankingsystem/Account.java
@@ -17,7 +17,7 @@ public class Account {
 		return balance;
 	}
 
-	public void increaseBalance(int amount) {
+	public void increaseBalance(long amount) {
 		balance += amount;
 	}
